refactor: extract navigation constants
